Updated for the 3.0 release
git-svn-id: https://svn.cv.nrao.edu/svn/casa/prerelease@9792 ac9c0da5-131f-0410-9009-a9c195f183e6

+++ casacore/casa/casa/Arrays/ArrayOpsDiffShapes.tcc
@@ -23,7 +23,7 @@
 //#                        520 Edgemont Road
 //#                        Charlottesville, VA 22903-2475 USA
 //#
-//# $Id$
+//# $Id: ArrayOpsDiffShapes.tcc 20705 2009-09-03 09:04:46Z gervandiepen $
 #include <casa/Arrays/ArrayMath.h>
 #include <casa/Arrays/ArrayLogical.h>
 //#include <casa/Arrays/ArrayUtil.h>
@@ -42,7 +42,7 @@ LogicalArray reformedMask(const Array<T>& data, const T truthvalue,
      && data.shape() == desiredform){
     return (data == truthvalue);
   }
-  else if(data.nelements() == desiredform.product()){
+  else if(static_cast<Int>(data.nelements()) == desiredform.product()){
     return (data == truthvalue).reform(desiredform);
   }
   else{
