+++ casacore/casa/casa/Arrays/Slicer.cc
@@ -23,7 +23,7 @@
 //#                        520 Edgemont Road
 //#                        Charlottesville, VA 22903-2475 USA
 //#
-//# $Id: Slicer.cc 18093 2004-11-30 17:51:10Z ddebonis $
+//# $Id: Slicer.cc 20699 2009-09-02 12:21:07Z gervandiepen $
 
 #include <casa/Arrays/Slicer.h>
 #include <casa/Arrays/Slice.h>
@@ -204,7 +204,8 @@ void Slicer::fillFixed()
     }
 }
 
-void Slicer::fillSlice (const Slice& sl, Int& start, Int& len, Int& stride)
+void Slicer::fillSlice (const Slice& sl, ssize_t& start, ssize_t& len,
+                        ssize_t& stride)
 {
     //# Initialized values will do in case an "entire" slice is given.
     if (!sl.all()) {
