+++ casacore/casa/casa/OS/HostInfoHpux.h
@@ -1,5 +1,5 @@
 //# HostInfo_hpux.h: HP/UX specific memory, swap, and CPU code.
-//# $Id: HostInfoHpux.h 18093 2004-11-30 17:51:10Z ddebonis $
+//# $Id: HostInfoHpux.h 20699 2009-09-02 12:21:07Z gervandiepen $
 
  /*
  **  This is a greatly MODIFIED version of a "top" machine dependent file.
@@ -81,13 +81,13 @@ friend class HostInfo;
     int valid;
     int cpus;
 
-    int swap_total;
-    int swap_used;
-    int swap_free;
+    ssize_t swap_total;
+    ssize_t swap_used;
+    ssize_t swap_free;
 
-    int memory_total;
-    int memory_used;
-    int memory_free;
+    ssize_t memory_total;
+    ssize_t memory_used;
+    ssize_t memory_free;
 
     int pageshift;		/* log base 2 of the pagesize */
 };
